requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.3
django-celery-results==2.4.0
django-environ==0.9.0
djangorestframework==3.14.0
kombu==5.2.4
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
